Added Refresh Tokens

+++ TweetBook.Api/Controllers/V1/PostsController.cs
@@ -5,6 +5,8 @@ using System;
 using System.Threading.Tasks;
 using TweetBook.Contracts.V1.Requests;
 using TweetBook.Contracts.V1.Responses;
+using TweetBook.Infrastructure.DTO;
+using TweetBook.Infrastructure.Extensions;
 using TweetBook.Infrastructure.Services;
 
 namespace TweetBook.Controllers.V1
@@ -18,7 +20,7 @@ namespace TweetBook.Controllers.V1
             _postService = postService;
         }
 
-
+        [Authorize]
         [HttpGet(TweetBook.Contracts.V1.ApiRoutes.Posts.Get)]
         public async Task<IActionResult> Get([FromRoute] Guid postId)
         {
@@ -33,56 +35,76 @@ namespace TweetBook.Controllers.V1
         }
 
 
-        
+        [Authorize]
         [HttpGet(TweetBook.Contracts.V1.ApiRoutes.Posts.GetAll)]
         public async Task<IActionResult> GetAll()
         {
             return Ok(await _postService.GetPostsAsync());
         }
 
+        [Authorize]
         [HttpPost(Contracts.V1.ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
+            var post = new PostDTO
+            {
+                //Id = Guid.NewGuid(),
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
 
 
 
-            var postId = Guid.NewGuid();
-
-            if (string.IsNullOrWhiteSpace(postRequest.Name))
-                postRequest.Name = "Example name";
-
-            await _postService.CreatePostAsync(postId, postRequest.Name);
+            await _postService.CreatePostAsync(post);
             var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.ToUriComponent()}";
-            var locationUri = baseUrl + "/" + Contracts.V1.ApiRoutes.Posts.Get.Replace("{postId}", postId.ToString());
+            var locationUri = baseUrl + "/" + Contracts.V1.ApiRoutes.Posts.Get.Replace("{postId}", post.Id.ToString());
 
             var response = new PostResponse
             {
-                Id = postId,
-                Name = postRequest.Name
+                Id = post.Id,
+                Name = post.Name
 
             };
 
             return Created(locationUri, response);
         }
 
+        [Authorize]
         [HttpPut(TweetBook.Contracts.V1.ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest postRequest)
         {
-            var updated = await _postService.UpdatePostAsync(postId, postRequest);
 
-            var response = new PostResponse
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if(!userOwnsPost)
             {
-                Id = postId,
-                Name = postRequest.Name
-            };
+                return BadRequest(new { error = "You do not own this post"});
+            }
+            var post = await _postService.GetPostByIdAsync(postId);
+
+            post.Name = postRequest.Name;
+
+            var updated = await _postService.UpdatePostAsync(post);
+
+            
+
             if (updated)
-                return Ok(response);
+                return Ok(post);
             return NotFound();
         }
 
+        [Authorize]
         [HttpDelete(Contracts.V1.ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this post" });
+            }
+
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
